Run tests with custom CPU/RAM resources limits
Some tests are failing in CI when default CPU/RAM scylla-ccm resources limits
are used.
Particularly the tests that perform node decommission are failing with
timeout as the node stuck when changing its state.

The change sets custom scylla-ccm resources limits for CPU and RAM -
smp=2 and memory=2G (which 1G per core, as the memory is divided
between cores).