Use --init flag in docker run command when running tests
Some ccm operations (e.g. node stop) are implemented as system sygnals.
But when tests are executed in a docker container in detached/non-interactive
mode (basically, the main flow of driver matrix execution), such signals are
not propagated properly to child processes, causing a few tests to fail.

This change helps to resolve signal handling problems by adding `--init` flag
to `docker run` command.
Note: with this change we no longer need to explicitly ignore the test classes,
which are annoted with skip flag in the driver code - setup/teardown execution
of these tests takes negligible amount of time after the fix.

Additionally, this changes adds patching of the
Should_AllMetricsHaveValidValues_When_NodeIsDown test, to use 3 nodes in the test
cluster, to ensure proper funtioning of raft algorithm.